update roles & permissions

+++ src/main/java/com/volunteer/main/repositories/PermissionRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PermissionRepository extends CrudRepository<PermissionEntity, Integer> {
+public interface PermissionRepository extends CrudRepository<PermissionEntity, Long> {
 
 }
